+++ src/backend/spreadsheet/Spreadsheet.h
@@ -89,7 +89,7 @@ public:
 	//data import
 	int prepareImport(std::vector<void*>& dataContainer, AbstractFileFilter::ImportMode,
 		int rows, int cols, QStringList colNameList, QVector<AbstractColumn::ColumnMode>) override;
-	void finalizeImport(int columnOffset, int startColumn , int endColumn,
+	void finalizeImport(size_t columnOffset, size_t startColumn , size_t endColumn,
 		const QString& dateTimeFormat, AbstractFileFilter::ImportMode) override;
 	int resize(AbstractFileFilter::ImportMode, QStringList colNameList, int cols);
